DSCHA_Harness/requirements.txt
django
requests
djangorestframework
celery
psycopg2
django-celery-beat
prctl
